requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
black==23.12.1
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
click==8.1.7
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
Django==4.2.8
django-multiupload==0.6.1
django-redis==5.4.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
mypy-extensions==1.0.0
numpy==1.24.4
packaging==23.2
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclean==2.7.6
pycparser==2.21
pyOpenSSL==23.3.0
redis==5.0.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
whitenoise==6.6.0
zope.interface==6.1
